requirements.txt

fastapi
uvicorn
python-telegram-bot==13.15
requests
